Refactor langchain (#10532)
